feat: update price offer queries to cast location IDs to bigint for consistency

+++ backend/services/booking/db/price_offers_query.sql.go
@@ -159,8 +159,8 @@ func (q *Queries) CreatePriceOffer(ctx context.Context, arg CreatePriceOfferPara
 const getPriceOfferById = `-- name: GetPriceOfferById :one
 SELECT price_offers.id, price_offers.token, price_offers.agent_id, price_offers.status, price_offers.name, price_offers.pickup_location_id, price_offers.dropoff_location_id, price_offers.pickup_date, price_offers.return_date, price_offers.pickup_time, price_offers.dropoff_time, price_offers.driver_age, price_offers.supplier_code, price_offers.car_details, price_offers.plan_inclusions, price_offers.currency_code, price_offers.purchase_price, price_offers.markup_percentage, price_offers.broker_erp_price, price_offers.bt_erp_price, price_offers.total_price, price_offers.offered_currency_code, price_offers.offered_price, price_offers.created_at, price_offers.updated_at , pl.name AS pickup_location, dl.name AS dropoff_location
 FROM price_offers
-    JOIN locations pl ON price_offers.pickup_location_id = pl.id
-    JOIN locations dl ON price_offers.dropoff_location_id = dl.id
+    JOIN locations pl ON price_offers.pickup_location_id::bigint = pl.id
+    JOIN locations dl ON price_offers.dropoff_location_id::bigint = dl.id
 WHERE price_offers.id = $1 AND price_offers.agent_id = $2
 `
 
@@ -237,8 +237,8 @@ func (q *Queries) GetPriceOfferById(ctx context.Context, arg GetPriceOfferByIdPa
 const getPriceOfferByToken = `-- name: GetPriceOfferByToken :one
 SELECT price_offers.id, price_offers.token, price_offers.agent_id, price_offers.status, price_offers.name, price_offers.pickup_location_id, price_offers.dropoff_location_id, price_offers.pickup_date, price_offers.return_date, price_offers.pickup_time, price_offers.dropoff_time, price_offers.driver_age, price_offers.supplier_code, price_offers.car_details, price_offers.plan_inclusions, price_offers.currency_code, price_offers.purchase_price, price_offers.markup_percentage, price_offers.broker_erp_price, price_offers.bt_erp_price, price_offers.total_price, price_offers.offered_currency_code, price_offers.offered_price, price_offers.created_at, price_offers.updated_at , pl.name AS pickup_location, dl.name AS dropoff_location
 FROM price_offers
-    JOIN locations pl ON price_offers.pickup_location_id = pl.id
-    JOIN locations dl ON price_offers.dropoff_location_id = dl.id
+    JOIN locations pl ON price_offers.pickup_location_id::bigint = pl.id
+    JOIN locations dl ON price_offers.dropoff_location_id::bigint = dl.id
  WHERE token = $1
 `
 
@@ -314,8 +314,8 @@ pickup_date, return_date, pickup_time, dropoff_time,
 currency_code, total_price, offered_currency_code, offered_price, 
 price_offers.created_at
 FROM price_offers
-    JOIN locations pl ON price_offers.pickup_location_id = pl.id
-    JOIN locations dl ON price_offers.dropoff_location_id = dl.id
+    JOIN locations pl ON price_offers.pickup_location_id::bigint = pl.id
+    JOIN locations dl ON price_offers.dropoff_location_id::bigint = dl.id
 WHERE agent_id = $1
   AND ($2::offer_status IS NULL OR status = $2::offer_status)
   AND ($3::text IS NULL OR price_offers.name ILIKE '%' || $3::text || '%')
